test_bitcoin: fix decorators for running tests twice
This bug is triggered if ecc_fast._libsecp256k1 and/or
crypto.AES are not present.

Before, if the first test would have raised an exception
it would have been implicitly caught by returning from
the finally block and hence this effectively masks the
test's outcome. Now, the exception is properly propagated
causing the test to fail if an exception is raised.

The effect of the return statement in a try block can be
observed using the following snippet.

def foo():
   try:
       raise Exception()
   finally:
       return

try:
    foo()
except:
    print("Exception?")
else:
    print("No exception!!!")